small fix of startNavigation() method to the abstract class Page.
